// packages/sdk/src/generated/pool/programs/splitRiskPool.ts
export enum SplitRiskPoolInstruction {
  AcceptGovernanceAuthority,
  AccrueFees,
  ActivateShielded,
  AddToken,
  AddTokenInitial,
  AdvanceProtectorEpoch,
  AllowAccount,
  CancelUnlock,
  ClaimCommission,
  ClaimRewards,
  ClosePool,
  CreatePool,
  DeactivateDustPool,
  DeactivatePool,
  DeactivateProtectorOnlyPool,
  DepositBacking,
  DepositShielded,
  DisallowAccount,
  FinalizeBootstrap,
  ForfeitCommission,
  InitFactory,
  PartialWithdrawProtector,
  PartialWithdrawShielded,
  Pause,
  PayPoolFee,
  PayProtocolFee,
  RemoveToken,
  SetGovernanceAuthority,
  SetMaxActivePools,
  SetMinimumCreationBondUsd,
  SetTokenRequiresStrictProtectedPrice,
  SettleExpiredProtectorPosition,
  StartUnlock,
  SweepInactiveProtectorBackingDust,
  SweepUnaccountedSurplus,
  Unpause,
  UpdateMinimumCollateral,
  UpdatePoolConfig,
  WithdrawProtector,
  WithdrawShielded,
}